+++ src/Core/Validator.cs
@@ -12,7 +12,7 @@ namespace NinjaCalc.Core {
     /// <summary>
     /// Encapsulates a validation result.
     /// </summary>
-    public class CalcValidationResult {
+    public class CalcValidationLevel {
 
         public string Name {
             get;
@@ -29,7 +29,7 @@ namespace NinjaCalc.Core {
             set;
         }
 
-        public CalcValidationResult(string name, System.Windows.Media.Brush borderBrush, System.Windows.Media.Brush backgroundBrush) {
+        public CalcValidationLevel(string name, System.Windows.Media.Brush borderBrush, System.Windows.Media.Brush backgroundBrush) {
             this.Name = name;
             this.BorderBrush = borderBrush;
             this.BackgroundBrush = backgroundBrush;
@@ -39,15 +39,35 @@ namespace NinjaCalc.Core {
     /// <summary>
     /// The different ValidationLevels a validation function can return.
     /// </summary>
-    public class CalcValidationResults {
-        public static readonly CalcValidationResult Ok;
-        public static readonly CalcValidationResult Warning;
-        public static readonly CalcValidationResult Error;
+    public class CalcValidationLevels {
+        public static readonly CalcValidationLevel Ok;
+        public static readonly CalcValidationLevel Warning;
+        public static readonly CalcValidationLevel Error;
 
-        static CalcValidationResults() {
-            Ok = new CalcValidationResult("ok", System.Windows.Media.Brushes.Green, (SolidColorBrush)new BrushConverter().ConvertFromString("#e5ffe5"));
-            Warning = new CalcValidationResult("warning", System.Windows.Media.Brushes.Orange, (SolidColorBrush)new BrushConverter().ConvertFromString("#fff5e5"));
-            Error = new CalcValidationResult("error", System.Windows.Media.Brushes.Red, (SolidColorBrush)new BrushConverter().ConvertFromString("#ffe5e5"));
+        static CalcValidationLevels() {
+            Ok = new CalcValidationLevel("ok", System.Windows.Media.Brushes.Green, (SolidColorBrush)new BrushConverter().ConvertFromString("#e5ffe5"));
+            Warning = new CalcValidationLevel("warning", System.Windows.Media.Brushes.Orange, (SolidColorBrush)new BrushConverter().ConvertFromString("#fff5e5"));
+            Error = new CalcValidationLevel("error", System.Windows.Media.Brushes.Red, (SolidColorBrush)new BrushConverter().ConvertFromString("#ffe5e5"));
+        }
+    }
+
+    /// <summary>
+    /// Designed to represent a single result from performing a validation on a calculator variable.
+    /// </summary>
+    public class CalcValidationResult {
+        public CalcValidationLevel CalcValidationLevel {
+            get;
+            set;
+        }
+
+        public string Message {
+            get;
+            set;
+        }
+
+        public CalcValidationResult(CalcValidationLevel calcValidationLevel, string message) {
+            this.CalcValidationLevel = calcValidationLevel;
+            this.Message = message;
         }
     }
 
@@ -59,7 +79,7 @@ namespace NinjaCalc.Core {
         /// <summary>
         /// Gets or sets the validation function which performs the validation and returns a ValidationResult_t.
         /// </summary>
-        public Func<double, CalcValidationResult> ValidationFunction {
+        public Func<double, CalcValidationLevel> ValidationFunction {
             get;
             set;
         }
@@ -77,7 +97,7 @@ namespace NinjaCalc.Core {
         /// Constructor. 
         /// </summary>
         /// <param name="validationFunction">The function which will validate the calculator variable.</param>
-        public Validator(Func<double, CalcValidationResult> validationFunction, string message) {
+        public Validator(Func<double, CalcValidationLevel> validationFunction, string message) {
             // Save function internally
             this.ValidationFunction = validationFunction;
 
@@ -91,7 +111,7 @@ namespace NinjaCalc.Core {
         /// it will return "CalcValidationResults.Ok".
         /// </summary>
         /// <returns>A validator which will give an error if the calculator variable is not a valid number.</returns>
-        public static Validator IsNumber(CalcValidationResult desiredValidationResult) {
+        public static Validator IsNumber(CalcValidationLevel desiredValidationResult) {
             return new Validator(
                 (value) => {
                     //return ValidationResult_t.Error;
@@ -99,7 +119,7 @@ namespace NinjaCalc.Core {
                         return desiredValidationResult;
                     }
                     else {
-                        return CalcValidationResults.Ok;
+                        return CalcValidationLevels.Ok;
                     }
                 },
                 "Value must be a real number.");
@@ -110,7 +130,7 @@ namespace NinjaCalc.Core {
         /// not greater than 0. If the number is greater than 0, it will return "CalcValidationResults.Ok".
         /// </summary>
         /// <returns>A validator which will give an error if the calculator variable is not a valid number.</returns>
-        public static Validator IsGreaterThanZero(CalcValidationResult desiredValidationResult) {
+        public static Validator IsGreaterThanZero(CalcValidationLevel desiredValidationResult) {
             return new Validator(
                 (value) => {
                     //return ValidationResult_t.Error;
@@ -118,7 +138,7 @@ namespace NinjaCalc.Core {
                         return desiredValidationResult;
                     }
                     else {
-                        return CalcValidationResults.Ok;
+                        return CalcValidationLevels.Ok;
                     }
                 },
                 "Value must be positive and not equal to 0.");