+++ changelog.md
@@ -3,6 +3,7 @@ v0.3.1
 
 - Added first unit tests using the inbuilt Microsoft Unit Testing framework.
 - Added IsGreaterThanZero() validator factory method, and added this constraint to all variables in the Ohm's Law calculator.
+- Tooltips on calculator variable textboxes are now displaying the results from value validation.
 
 v0.3.0
 ------